Cleaning up the main endpoint function that express uses to route requests.
